cli: Move formatters to a separate package #696
