+++ cli/zally/utils/formatters/markdown_formatter.go
@@ -1,4 +1,4 @@
-package utils
+package formatters
 
 import (
 	"bytes"
