+++ cli/zally/utils/formatters/markdown_formatter_test.go
@@ -1,4 +1,4 @@
-package utils
+package formatters
 
 import (
 	"fmt"
