+++ cli/zally/utils/formatters/pretty_colorizer_test.go
@@ -1,4 +1,4 @@
-package utils
+package formatters
 
 import (
 	"fmt"
