+++ cli/zally/utils/formatters/pretty_formatter.go
@@ -1,4 +1,4 @@
-package utils
+package formatters
 
 import (
 	"bytes"
